+++ docsrc/conf.py
@@ -11,6 +11,10 @@ repo_name = 'elements'
 
 setup_globals(org, org_name, repo, repo_name)
 
-graphviz_output_format = 'svg'
-
 html_theme = "foundation"
+
+intersphinx_mapping = {
+    'python': ('https://docs.python.org/3', None),
+    'sympy': ('https://docs.sympy.org/latest', None),
+    'euclid': ('https://geometor.github.io/euclid', None),
+}